update mentorship card to use the new tags
